chore: move PricingPage and ContactPage to src/pages/
Resolves #81
